(svn r23138) -Feature: [NewGRF] Allow passing 32bit parameters to 60+x variables (using var 7B). Currently most useful for vehicle var 60.

+++ src/newgrf_airport.cpp
@@ -127,7 +127,7 @@ void AirportOverrideManager::SetEntitySpec(AirportSpec *as)
 	}
 }
 
-uint32 AirportGetVariable(const ResolverObject *object, byte variable, byte parameter, bool *available)
+uint32 AirportGetVariable(const ResolverObject *object, byte variable, uint32 parameter, bool *available)
 {
 	const Station *st = object->u.airport.st;
 	byte layout       = object->u.airport.layout;
